Allow for the Dirus A40 using 1-based addresses.
The A40 should be receiving 0-indexed addresses in the modbus PDU, and
then adding 1 to them to turn them into the addresses specified in the
manual.  But either it isn't, or the manual has listed the addresses
from the user's (request-construction) point of view (which is quite
sensible really).

This commit fixes up two things:

1. The request wrapper in jem_data makes 1-based requests, rather than
   0-based requests in order that it is compatible with the A40.

2. The server simulator now behaves like an A40 by accepting 1-based
   requests.